Extract text document retrival logic from ActiveDocumentLocator
