+++ src/Integration.Vsix.UnitTests/Helpers/ActiveDocumentLocatorTests.cs
@@ -20,11 +20,9 @@
 
 using FluentAssertions;
 using Microsoft.VisualStudio;
-using Microsoft.VisualStudio.Editor;
 using Microsoft.VisualStudio.Shell.Interop;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Microsoft.VisualStudio.Text;
-using Microsoft.VisualStudio.TextManager.Interop;
 using Moq;
 using SonarLint.VisualStudio.Integration.Vsix;
 
@@ -34,12 +32,8 @@ namespace SonarLint.VisualStudio.Integration.UnitTests.Helpers
     public class ActiveDocumentLocatorTests
     {
         private Mock<IVsMonitorSelection> monitorSelectionMock;
-        private Mock<IVsEditorAdaptersFactoryService> adapterServiceMock;
-        private Mock<IVsWindowFrame> frameMock;
-        private Mock<IVsTextBuffer> vsTextBufferMock;
-
-        private IVsWindowFrame ValidVsWindowFrame => frameMock.Object;
-        private IVsTextBuffer ValidVsTextBuffer => vsTextBufferMock.Object;
+        private Mock<ITextDocumentProvider> textDocumentProviderMock;
+        private ITextDocument textDocument;
 
         private ActiveDocumentLocator testSubject;
 
@@ -49,12 +43,10 @@ namespace SonarLint.VisualStudio.Integration.UnitTests.Helpers
             ThreadHelper.SetCurrentThreadAsUIThread();
 
             monitorSelectionMock = new Mock<IVsMonitorSelection>();
-            adapterServiceMock = new Mock<IVsEditorAdaptersFactoryService>();
+            textDocumentProviderMock = new Mock<ITextDocumentProvider>();
+            textDocument = Mock.Of<ITextDocument>();
 
-            testSubject = new ActiveDocumentLocator(monitorSelectionMock.Object, adapterServiceMock.Object);
-
-            frameMock = new Mock<IVsWindowFrame>();
-            vsTextBufferMock = new Mock<IVsTextBuffer>();
+            testSubject = new ActiveDocumentLocator(monitorSelectionMock.Object, textDocumentProviderMock.Object);
         }
 
         [TestMethod]
@@ -72,7 +64,7 @@ namespace SonarLint.VisualStudio.Integration.UnitTests.Helpers
         }
 
         [TestMethod]
-        public void Find_OpenDocument_NotanIVsWindowFrame()
+        public void Find_OpenDocument_NotAnIVsWindowFrame()
         {
             // Arrange
             var notAWindowFrame = new object();
@@ -87,139 +79,28 @@ namespace SonarLint.VisualStudio.Integration.UnitTests.Helpers
         }
 
         [TestMethod]
-        public void Find_OpenDocument_FrameDoesNotHaveDocData()
+        public void Find_OpenDocument_IVsWindowFrame_ReturnsDocument()
         {
             // Arrange
-            Configure(
-                activeFrame: ValidVsWindowFrame,
-                activeFrameDocDataObject: null);
+            var windowFrame = Mock.Of<IVsWindowFrame>();
+            Configure(activeFrame: windowFrame);
 
             // Act
             var result = testSubject.FindActiveDocument();
 
             // Assert
-            result.Should().BeNull();
-            VerifyMonitorSelectionServiceCalled(ValidVsWindowFrame);
-            VerifyFrameGetPropertyCalled(null);
+            result.Should().Be(textDocument);
+            VerifyMonitorSelectionServiceCalled(windowFrame);
         }
 
-        [TestMethod]
-        public void Find_OpenDocument_FrameDocDataIsNotVsTextBuffer()
-        {
-            // Arrange
-            var notVsTextBuffer = new object();
-
-            Configure(
-                activeFrame: ValidVsWindowFrame,
-                activeFrameDocDataObject: notVsTextBuffer);
-
-            // Act
-            var result = testSubject.FindActiveDocument();
-
-            // Assert
-            result.Should().BeNull();
-            VerifyMonitorSelectionServiceCalled(ValidVsWindowFrame);
-            VerifyFrameGetPropertyCalled(notVsTextBuffer);
-            VerifyAdapterServiceNotCalled();
-        }
-
-        [TestMethod]
-        public void Find_OpenDocument_FrameDocDataIsTextBuffer_AdapterServiceReturnsNull()
-        {
-            // Arrange            
-            Configure(
-                activeFrame: ValidVsWindowFrame,
-                activeFrameDocDataObject: ValidVsTextBuffer);
-
-            // Act
-            var result = testSubject.FindActiveDocument();
-
-            // Assert
-            result.Should().BeNull();
-            VerifyMonitorSelectionServiceCalled(ValidVsWindowFrame);
-            VerifyFrameGetPropertyCalled(ValidVsTextBuffer);
-            VerifyAdapterServiceCalled(ValidVsTextBuffer);
-        }
-
-
-        [TestMethod]
-        public void Find_OpenDocument_NewTextBuffer_NoTextDoc()
-        {
-            // Arrange
-            var newTextBuffer = CreateValidMefTextBuffer(null);
-
-            Configure(
-                activeFrame: ValidVsWindowFrame,
-                activeFrameDocDataObject: ValidVsTextBuffer,
-                adapterServiceConvertedTextBuffer: newTextBuffer);
-
-            // Act
-            var result = testSubject.FindActiveDocument();
-
-            // Assert
-            result.Should().Be(null);
-
-            VerifyMonitorSelectionServiceCalled(ValidVsWindowFrame);
-            VerifyFrameGetPropertyCalled(ValidVsTextBuffer);
-            VerifyAdapterServiceCalled(ValidVsTextBuffer);
-        }
-
-
-        [TestMethod]
-        public void Find_OpenDocument_NewTextBuffer_HasTextDoc()
-        {
-            // Arrange
-            var newTextDoc = new Mock<ITextDocument>().Object;
-            var newTextBuffer = CreateValidMefTextBuffer(newTextDoc);
-
-            Configure(
-                activeFrame: ValidVsWindowFrame,
-                activeFrameDocDataObject: ValidVsTextBuffer,
-                adapterServiceConvertedTextBuffer: newTextBuffer);
-
-            // Act
-            var result = testSubject.FindActiveDocument();
-
-            // Assert
-            result.Should().Be(newTextDoc);
-
-            VerifyMonitorSelectionServiceCalled(ValidVsWindowFrame);
-            VerifyFrameGetPropertyCalled(ValidVsTextBuffer);
-            VerifyAdapterServiceCalled(ValidVsTextBuffer);
-        }
-
-        private void Configure(
-            object activeFrame = null,
-            object activeFrameDocDataObject = null,
-            ITextBuffer adapterServiceConvertedTextBuffer = null
-            )
+        private void Configure(object activeFrame = null)
         {
             monitorSelectionMock.Setup(x => x.GetCurrentElementValue((uint)VSConstants.VSSELELEMID.SEID_DocumentFrame, out activeFrame));
-            frameMock.Setup(x => x.GetProperty((int)__VSFPROPID.VSFPROPID_DocData, out activeFrameDocDataObject));
-            adapterServiceMock.Setup(x => x.GetDocumentBuffer(activeFrameDocDataObject as IVsTextBuffer)).Returns(adapterServiceConvertedTextBuffer);
-        }
-
-        private ITextBuffer CreateValidMefTextBuffer(ITextDocument textDocument)
-        {
-            var newTextBufferMock = new Mock<ITextBuffer>();
-
-            var bufferProps = new Microsoft.VisualStudio.Utilities.PropertyCollection();
-            bufferProps.AddProperty(typeof(ITextDocument), textDocument);
-            newTextBufferMock.Setup(x => x.Properties).Returns(bufferProps);
-
-            return newTextBufferMock.Object;
+            var vsFrame = activeFrame as IVsWindowFrame;
+            textDocumentProviderMock.Setup(x => x.GetFromFrame(vsFrame)).Returns(textDocument);
         }
 
         private void VerifyMonitorSelectionServiceCalled(object obj) =>
             monitorSelectionMock.Verify(x => x.GetCurrentElementValue((uint)VSConstants.VSSELELEMID.SEID_DocumentFrame, out obj), Times.Once);
-
-        private void VerifyFrameGetPropertyCalled(object obj) =>
-            frameMock.Verify(x => x.GetProperty((int)__VSFPROPID.VSFPROPID_DocData, out obj),Times.Once);
-
-        private void VerifyAdapterServiceCalled(IVsTextBuffer inputVsBuffer) =>
-            adapterServiceMock.Verify(x => x.GetDocumentBuffer(inputVsBuffer), Times.Once);
-
-        private void VerifyAdapterServiceNotCalled() =>
-            adapterServiceMock.Verify(x => x.GetDocumentBuffer(It.IsAny<IVsTextBuffer>()), Times.Never);
     }
 }
