+++ src/Integration.Vsix/CFamily/CFamilyReproducerCommand.cs
@@ -22,7 +22,6 @@ using System;
 using System.ComponentModel.Design;
 using System.Diagnostics;
 using System.Linq;
-using Microsoft.VisualStudio.Editor;
 using Microsoft.VisualStudio.Shell;
 using Microsoft.VisualStudio.Shell.Interop;
 using SonarLint.VisualStudio.Core;
@@ -58,8 +57,8 @@ namespace SonarLint.VisualStudio.Integration.Vsix.CFamily
             await ThreadHelper.JoinableTaskFactory.SwitchToMainThreadAsync(package.DisposalToken);
 
             var monitorSelection = await package.GetServiceAsync(typeof(SVsShellMonitorSelection)) as IVsMonitorSelection;
-            var adapterService = await package.GetMefServiceAsync<IVsEditorAdaptersFactoryService>();
-            var docLocator = new ActiveDocumentLocator(monitorSelection, adapterService);
+            var textDocumentProvider = await package.GetMefServiceAsync<ITextDocumentProvider>();
+            var docLocator = new ActiveDocumentLocator(monitorSelection, textDocumentProvider);
 
             var languageRecognizer = await package.GetMefServiceAsync<ISonarLanguageRecognizer>();
             var requester = await package.GetMefServiceAsync<IAnalysisRequester>();
